- Split up theme_provider, themecolor_provider and thememode_provider for better readability/usability.
